requirements.txt
Django==1.5

h5py==2.1.3
numpy==1.7.1
scipy==0.12.0
rpy2==2.3.6

python-memcached==1.53
httplib2==0.8
python-openid==2.2.5
recaptcha-client==1.0.6
wsgiref==0.1.2
markdown>=2.1

-e git://github.com/open-machine-learning/mldata-utils.git@e683ca4c536d6a74f9fd734b61f8e671a5527fa2#egg=mldata_utils-dev

